Refactoring resources
