+++ src/main/java/br/com/ceps/cepsbank/infrastructure/mappers/AccountMapper.java
@@ -1,9 +1,11 @@
 package br.com.ceps.cepsbank.infrastructure.mappers;
 
-import br.com.ceps.cepsbank.application.web.request.AccountRequest;
+import br.com.ceps.cepsbank.application.web.request.AccountEventRequest;
 import br.com.ceps.cepsbank.application.web.response.AccountResponse;
 import br.com.ceps.cepsbank.domain.account.model.Account;
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.Mappings;
 import org.mapstruct.factory.Mappers;
 
 @Mapper
@@ -11,7 +13,11 @@ public interface AccountMapper {
 
     AccountMapper MAPPER = Mappers.getMapper(AccountMapper.class);
 
-    Account toDomain(AccountRequest request);
+    @Mappings({
+            @Mapping(target = "origin.balance", source = "origin"),
+            @Mapping(target = "destination.balance", source = "destination")
+    })
+    Account toDomain(AccountEventRequest request);
     AccountResponse toResponse(Account domain);
 
 }
