+++ src/main/resources/application.yml
@@ -1,2 +1,1 @@
-server:
-  context-path: /ceps-bank/api/v1/
+
